requirements.txt
httpx
litellm
python-dotenv
weave-python
fastapi>=0.110.0
uvicorn>=0.27.0
jinja2>=3.1.3 